# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "Complete API for mechanic shop management with inventory tracking, JWT authentication, rate limiting, and caching"
  version: "1.0.0"
  contact:
    name: "Mechanic Shop API Support"
    email: "support@mechanicshop.com"
host: "be1-mechanic-shop-assignment.onrender.com"
schemes:
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
      JWT Authorization header using the Bearer scheme.
      
      **Format:** `Bearer {jwt_token}`
      
      **Example:** `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
      
      **How to obtain token:**
      1. Make a POST request to `/login` with valid credentials
      2. Copy the `access_token` from the response
      3. Add it to the Authorization header as `Bearer {access_token}`
      
      **Token contains:**
      - User identity and role information
      - Expiration time (typically 24 hours)
      - Permissions and access level
      
      **Usage:**
      - Include this header in requests to protected endpoints
      - Token will be validated on each request
      - Expired tokens will return 401 Unauthorized
  
  JWTAuth:
    type: apiKey
    name: Authorization
    in: header
    description: |
      JSON Web Token (JWT) authentication for API access.
      
      **Authentication Flow:**
      1. **Login:** POST credentials to `/login` endpoint
      2. **Token:** Receive JWT access token in response
      3. **Header:** Include token in Authorization header
      4. **Access:** Make authenticated requests to protected resources
      
      **Token Security:**
...
paths:
...
  /mechanics/:
    get:
      tags:
        - "Mechanics"
      summary: "Get all mechanics"
      description: "Get all mechanics (Cached for 10 minutes)"
      responses:
        200:
          description: "Mechanics retrieved successfully"
    post:
      tags:
        - "Mechanics"
      summary: "Create mechanic"
      description: "Create a new mechanic (Rate Limited: 10 per minute, Requires JWT)"
      security:
        - Bearer: []
      parameters:
        - in: body
          name: mechanic
          required: true
          schema:
            $ref: "#/definitions/MechanicCreate"
      responses:
        201:
          description: "Mechanic created successfully"
        400:
          description: "Validation error"
        401:
          description: "Unauthorized"
        429:
          description: "Rate limit exceeded"
  /mechanics/{id}:
    get:
      tags:
        - "Mechanics"
      summary: "Get mechanic by ID"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
      responses:
        200:
          description: "Mechanic retrieved successfully"
        404:
          description: "Mechanic not found"
    put:
      tags:
        - "Mechanics"
      summary: "Update mechanic"
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: integer
          required: true
        - in: body
          name: mechanic
          schema:
            $ref: "#/definitions/MechanicUpdate"
      responses:
        200:
          description: "Mechanic updated successfully"
        401:
          description: "Unauthorized"
        404:
          description: "Mechanic not found"
...
  /inventory/:
    get:
      tags:
        - "Inventory"
      summary: "Get all inventory"
      description: "Get all inventory with pagination, sorting, and filtering (Cached for 10 minutes)"
      parameters:
        - in: query
          name: page
          type: integer
          description: "Page number"
          default: 1
        - in: query
          name: per_page
          type: integer
          description: "Items per page (max 100)"
          default: 10
        - in: query
          name: sort_by
          type: string
          description: "Sort field"
          enum: ["id", "name", "price"]
        - in: query
          name: sort_order
          type: string
          description: "Sort order"
          enum: ["asc", "desc"]
        - in: query
          name: name
          type: string
          description: "Search by name (partial match)"
        - in: query
          name: min_price
          type: number
          description: "Minimum price filter"
        - in: query
          name: max_price
          type: number
          description: "Maximum price filter"
      responses:
        200:
          description: "Inventory retrieved successfully"
    post:
      tags:
        - "Inventory"
      summary: "Create inventory item"
      description: "Create a new inventory item (Requires JWT)"
      security:
        - Bearer: []
      parameters:
        - in: body
          name: inventory
          required: true
          schema:
            $ref: "#/definitions/InventoryCreate"
      responses:
        201:
          description: "Inventory item created successfully"
        400:
          description: "Validation error"
        401:
          description: "Unauthorized"
        409:
          description: "Item with same name already exists"
  /inventory/bulk:
    post:
      tags:
        - "Inventory"
      summary: "Bulk create inventory"
      description: "Create multiple inventory items at once (Requires JWT)"
      security:
        - Bearer: []
      parameters:
        - in: body
          name: inventory_list
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/InventoryCreate"
      responses:
        201:
          description: "Inventory items created successfully"
        400:
          description: "Validation error"
        401:
          description: "Unauthorized"
  /inventory/{id}:
    get:
      tags:
        - "Inventory"
      summary: "Get inventory item by ID"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
      responses:
        200:
          description: "Inventory item retrieved successfully"
        404:
          description: "Inventory item not found"
    put:
      tags:
        - "Inventory"
      summary: "Update inventory item"
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: integer
          required: true
        - in: body
          name: inventory
          schema:
            $ref: "#/definitions/InventoryUpdate"
      responses:
        200:
          description: "Inventory item updated successfully"
        401:
          description: "Unauthorized"
        404:
          description: "Inventory item not found"
        409:
          description: "Item with same name already exists"
    delete:
      tags:
        - "Inventory"
      summary: "Delete inventory item"
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: integer
          required: true
      responses:
        200:
          description: "Inventory item deleted successfully"
        401:
          description: "Unauthorized"
        404:
          description: "Inventory item not found"
...
  /service-tickets/{id}/inventory:
    post:
      tags:
        - "Service Tickets"
      summary: "Add inventory to ticket"
      description: "Add inventory parts to a service ticket"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
        - in: body
          name: inventory_data
          required: true
          schema:
            type: object
            properties:
              inventory_ids:
                type: array
                items:
                  type: integer
                example: [1, 2, 3]
      responses:
        200:
          description: "Inventory added to ticket successfully"
        400:
          description: "Validation error"
        404:
          description: "Service ticket not found"
    delete:
      tags:
        - "Service Tickets"
      summary: "Remove inventory from ticket"
      description: "Remove inventory parts from a service ticket"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
        - in: body
          name: inventory_data
          required: true
          schema:
            type: object
            properties:
              inventory_ids:
                type: array
                items:
                  type: integer
                example: [1, 2]
      responses:
        200:
          description: "Inventory removed from ticket successfully"
        400:
          description: "Validation error"
        404:
          description: "Service ticket not found"
...
definitions:
...
  MechanicCreate:
    type: object
    required:
      - name
      - specialty
      - hourly_rate
    properties:
      name:
        type: string
        example: "Mike Johnson"
      specialty:
        type: string
        example: "Engine Repair"
      hourly_rate:
        type: number
        format: float
        minimum: 0
        example: 75.00
  MechanicUpdate:
    type: object
    properties:
      name:
        type: string
        example: "Mike Johnson"
      specialty:
        type: string
        example: "Engine Repair"
      hourly_rate:
        type: number
        format: float
        example: 80.00
...
  InventoryCreate:
    type: object
    required:
      - name
      - price
    properties:
      name:
        type: string
        example: "Oil Filter"
      price:
        type: number
        format: float
        minimum: 0
        example: 15.99
  InventoryUpdate:
    type: object
    properties:
      name:
        type: string
        example: "Oil Filter"
      price:
        type: number
        format: float
        example: 18.99
...